# agent-team/engineering/docs/openapi-draft.yaml
openapi: 3.1.0
info:
  title: AI Multi-Agent Console API Draft
  version: 0.1.0
  description: Draft API contract for AI multi-agent web console M1.
servers:
  - url: /api
paths:
  /agents:
    get:
      summary: List agents
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [online, offline, busy, idle, degraded]
        - in: query
          name: keyword
          schema:
            type: string
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentSummary'
  /agents/{agentId}:
    get:
      summary: Get agent detail
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AgentDetail'
...
components:
  schemas:
    AgentSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        profileName: { type: string }
        status:
          type: string
          enum: [online, offline, busy, idle, degraded]
        currentSessionId:
          type: [string, 'null']
        lastSeenAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
        capabilities:
          type: array
          items: { type: string }
      required: [id, name, role, profileName, status, lastSeenAt, unreadCount, capabilities]
    AgentDetail:
      allOf:
        - $ref: '#/components/schemas/AgentSummary'
        - type: object
          properties:
            profilePath: { type: string }
            summary: { type: string }
          required: [profilePath, summary]
...